+++ NanoPlayBoard/src/LDR.h
@@ -15,7 +15,7 @@ class LDR
     int read();
 
   private:
-    int _pin;
+    uint8_t _pin;
 };
 
 #endif
